Use correct Stylecop, more comments, cleanup
