+++ csharp_dotnetcore/11.QnAMaker/AspNetCore-QnA-Bot.bot
@@ -4,12 +4,12 @@
     "secretKey": "",
     "services": [
       {
-        "id": "http://localhost:54000/api/messages",
+        "id": "http://localhost:54011/api/messages",
         "name": "AspNetCore-QnA-Bot",
         "type": "endpoint",
         "appId": "",
         "appPassword": "",
-        "endpoint": "http://localhost:54000/api/messages"
+        "endpoint": "http://localhost:54011/api/messages"
       },
       {
         "type": "qna",
