+++ csharp_dotnetcore/11.QnAMaker/Program.cs
@@ -1,11 +1,11 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-using Microsoft.AspNetCore;
-using Microsoft.AspNetCore.Hosting;
-
 namespace AspNetCore_QnA_Bot
 {
+    using Microsoft.AspNetCore;
+    using Microsoft.AspNetCore.Hosting;
+
     public class Program
     {
         public static void Main(string[] args)
