+++ csharp_dotnetcore/11.QnAMaker/Properties/launchSettings.json
@@ -3,7 +3,7 @@
     "windowsAuthentication": false,
     "anonymousAuthentication": true,
     "iisExpress": {
-      "applicationUrl": "http://localhost:54955/",
+      "applicationUrl": "http://localhost:54011/",
       "sslPort": 0
     }
   },
@@ -21,7 +21,7 @@
       "environmentVariables": {
         "ASPNETCORE_ENVIRONMENT": "Development"
       },
-      "applicationUrl": "http://localhost:54956/"
+      "applicationUrl": "http://localhost:54011/"
     }
   }
 }
