+++ csharp_dotnetcore/11.QnAMaker/README.md
@@ -14,20 +14,21 @@ In this sample, we demonstrate how to use the QnA Maker service to answer questi
 git clone https://github.com/Microsoft/botbuilder-samples.git
 ```
 
-
 ### Prerequisites
 - Follow instructions [here](https://docs.microsoft.com/en-us/azure/cognitive-services/qnamaker/how-to/set-up-qnamaker-service-azure) to create a QnA Maker service.
 - Follow instructions [here](https://docs.microsoft.com/en-us/azure/cognitive-services/qnamaker/how-to/migrate-knowledge-base) to import the [smartLightFAQ.tsv](smartLightFAQ.tsv) to your newly created QnA Maker service.
 - Update [appsettings.json](appsettings.json) with your QnAMaker-Host, QnAMaker-KnowledgeBaseId and QnAMaker-EndpointKey. You can find this information under "Settings" tab for your QnA Maker Knowledge Base at [QnAMaker.ai](https://qnamaker.ai)
 - Follow instructions [here](https://docs.microsoft.com/en-us/azure/application-insights/app-insights-asp-net-core) to set up your Application Insights service.
   - Note: The Application Insights will automatically update the [appsettings.json](appsettings.json) file.
+- (Optional) Follow instructions [here](https://github.com/Microsoft/botbuilder-tools/tree/master/packages/QnAMaker) to set up the Qna Maker CLI to deploy the model.
+
 
 ## Visual studio
-- Navigate to the samples folder (BotBuilder-Samples\11.QnAMaker) and open AspNetCore-QnA-Bot.csproj in Visual studio 
+- Navigate to the samples folder (`BotBuilder-Samples\csharp_dotnetcore\11.QnAMaker`) and open AspNetCore-QnA-Bot.csproj in Visual studio 
 - Hit F5
 
 ## Visual studio code
-- Open BotBuilder-Samples\11.QnAMaker sample folder.
+- Open `BotBuilder-Samples\csharp_dotnetcore\11.QnAMaker` sample folder.
 - Bring up a terminal, navigate to BotBuilder-Samples\11.QnAMaker folder
 - type 'dotnet run'
 
@@ -38,17 +39,12 @@ git clone https://github.com/Microsoft/botbuilder-samples.git
 
 ### Connect to bot using Bot Framework Emulator **V4**
 - Launch Bot Framework Emulator
-- File -> Open bot and navigate to BotBuilder-Samples\11.QnAMaker folder
+- File -> Open bot and navigate to `BotBuilder-Samples\csharp_dotnetcore\11.QnAMaker` folder
 - Select AspNetCore-QnA-Bot.bot file
-
-
-# QnA Maker service
-QnA Maker enables you to power a question and answer service from your semi-structured content. 
-
-One of the basic requirements in writing your own Bot service is to seed it with questions and answers. In many cases, the questions and answers already exist in content like FAQ URLs/documents, product manuals, etc. With QnA Maker, users can query your application in a natural, conversational manner. QnA Maker uses machine learning to extract relevant question-answer pairs from your content. It also uses powerful matching and ranking algorithms to provide the best possible match between the user query and the questions.
 
 # Further reading
 
 - [Azure Bot Service Introduction](https://docs.microsoft.com/en-us/azure/bot-service/bot-service-overview-introduction?view=azure-bot-service-4.0)
 - [Bot State](https://docs.microsoft.com/en-us/azure/bot-service/bot-builder-storage-concept?view=azure-bot-service-4.0)
 - [QnA Maker documentation](https://docs.microsoft.com/en-us/azure/cognitive-services/qnamaker/overview/overview)
+- [QnA Maker command line tool](https://github.com/Microsoft/botbuilder-tools/tree/master/packages/QnAMaker)
