+++ csharp_dotnetcore/11.QnAMaker/Startup.cs
@@ -1,6 +1,7 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
+using System;
 using AspNetCore_QnA_Bot.AppInsights;
 using Microsoft.ApplicationInsights;
 using Microsoft.ApplicationInsights.Extensibility;
@@ -12,14 +13,20 @@ using Microsoft.Bot.Builder.Integration.AspNet.Core;
 using Microsoft.Bot.Configuration;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
-using System;
 
 namespace AspNetCore_QnA_Bot
 {
+    /// <summary>
+    /// The Startup class configures services and the app's request pipeline.
+    /// </summary>
     public class Startup
     {
-        // This method gets called by the runtime. Use this method to add services to the container.
-        // For more information on how to configure your application, visit https://go.microsoft.com/fwlink/?LinkID=398940
+        /// <summary>
+        /// Initializes a new instance of the <see cref="Startup"/> class.
+        /// This method gets called by the runtime. Use this method to add services to the container.
+        /// For more information on how to configure your application, visit https://go.microsoft.com/fwlink/?LinkID=398940.
+        /// </summary>
+        /// <param name="env">Provides information about the web hosting environment an application is running in.</param>
         public Startup(IHostingEnvironment env)
         {
             var builder = new ConfigurationBuilder()
@@ -31,12 +38,21 @@ namespace AspNetCore_QnA_Bot
             Configuration = builder.Build();
         }
 
+        /// <summary>
+        /// Gets the configuration that represents a set of key/value application configuration properties.
+        /// </summary>
+        /// <value>
+        /// The IConfiguration that represents a set of key/value application configuration properties.
+        /// </value>
         public IConfiguration Configuration { get; }
 
-        // This method gets called by the runtime. Use this method to add services to the container.
+        /// <summary>
+        /// This method gets called by the runtime. Use this method to add services to the container.
+        /// </summary>
+        /// <param name="services">Specifies the contract for a collection of service descriptors.</param>
         public void ConfigureServices(IServiceCollection services)
         {
-            // Load Connected Services from .bot file
+            // Load the connected services from .bot file
             // BUGBUG: This needs to change to common bot file, once emulator understand appInsights type.
             var botConfig = BotConfiguration.LoadAsync(@".\QnaBot.bot").Result;
 
@@ -54,11 +70,34 @@ namespace AspNetCore_QnA_Bot
                 var appInsightsLogger = new MyAppInsightsLoggerMiddleware(connectedServices.TelemetryClient.InstrumentationKey, logUserName: true, logOriginalMessage: true);
                 options.Middleware.Add(appInsightsLogger);
             });
-            
-
         }
 
-        // Configure Bot Services
+        /// <summary>
+        /// This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
+        /// </summary>
+        /// <param name="app">The application builder.  This provides the mechanisms to configure an application's request pipeline.</param>
+        /// <param name="env">Provides information about the web hosting environment an application is running in.</param>
+        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        {
+            if (env.IsDevelopment())
+            {
+                app.UseDeveloperExceptionPage();
+            }
+
+            app.UseDefaultFiles()
+                .UseStaticFiles()
+                .UseBotFramework();
+        }
+
+        /// <summary>
+        /// Initialize the bot's references to external services.
+        ///
+        /// For example, Application Insights and QnaMaker services
+        /// are created here.  These external services are configured
+        /// using the BotConfigure class (based on the contents of your ".bot" file).
+        /// </summary>
+        /// <param name="config">Configuration object based on your ".bot" file.</param>
+        /// <returns>A object representing client objects to access external services the bot uses.</returns>
         private static BotServices InitBotServices(BotConfiguration config)
         {
             var connectedServices = new BotServices();
@@ -70,8 +109,10 @@ namespace AspNetCore_QnA_Bot
                     case ServiceTypes.QnA:
                         {
                             // Create a QNA Maker that is initialized and suitable for passing
-                            // into the IBot-derived class (QnABot). 
-                            // This custom class also logs results to Application Insights.
+                            // into the IBot-derived class (QnABot).
+                            // In this case, we're creating a custom class (wrapping the original
+                            // QnAMaker client) that logs the results of QnA Maker into Application
+                            // Insights for future anaysis.
                             var qna = service as QnAMakerService;
                             if (string.IsNullOrWhiteSpace(qna.KbId))
                             {
@@ -88,19 +129,14 @@ namespace AspNetCore_QnA_Bot
                                 throw new InvalidOperationException("The Qna Host ('hostname') is required to run this sample.  Please update your QnaBot.bot file.");
                             }
 
-                            
                             var qnaEndpoint = new QnAMakerEndpoint()
                             {
                                 KnowledgeBaseId = qna.KbId,
                                 EndpointKey = qna.EndpointKey,
                                 Host = qna.Hostname,
                             };
-                            var qnaOptions = new QnAMakerOptions()
-                            {
-                                ScoreThreshold = 0.3f
-                            };
 
-                            var qnaMaker = new MyAppInsightsQnaMaker(qnaEndpoint, qnaOptions, logUserName: false, logOriginalMessage: false);
+                            var qnaMaker = new MyAppInsightsQnaMaker(qnaEndpoint, null, logUserName: false, logOriginalMessage: false);
                             connectedServices.QnAServices.Add(qna.Name, qnaMaker);
 
                             break;
@@ -121,22 +157,8 @@ namespace AspNetCore_QnA_Bot
                         }
                 }
             }
+
             return connectedServices;
-        }
-
-
-
-        // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
-        {
-            if (env.IsDevelopment())
-            {
-                app.UseDeveloperExceptionPage();
-            }
-
-            app.UseDefaultFiles()
-                .UseStaticFiles()
-                .UseBotFramework();
         }
     }
 }
